pyproject.toml
[project]
name = "bilibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "bilibili-api-python>=17.4.0",
    "python-dotenv>=1.2.1",
    "textual[syntax]>=6.8.0",
    "textual-dev>=1.8.0",
    "ollama>=0.6.1",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
]
